namazing-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "namazing"
version = "0.1.0"
description = "AI-powered baby naming service"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Namazing Team" }
]
dependencies = [
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
namazing = "namazing.cli.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/namazing"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
